FIX: [9.4_RC] plugin image_text - check file exists [t25583]
git-svn-id: http://svn.resourcespace.com/svn/rs/releases/9.4_RC@16182 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ plugins/image_text/hooks/collection_download.php
@@ -99,7 +99,10 @@ function HookImage_textCollection_downloadModifydownloadfile()
 		$p=$newdlfile;
 		}
 		
-	unlink($tmpolfile);
+    if(file_exists($tmpolfile))
+        {
+        unlink($tmpolfile);
+        }  
 	
 	return true;
 	}
